+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Latest Release: 1.1.0
 
 * Added Fallback Image Support
+* Added Cross Browser Window Resize Event
 
 ### 1.0.0